requirements.txt
openai==0.28.1
pyreadr==0.4.4
vllm==0.5.0
python-dev-tools 
seaborn
matplotlib
pandas
numpy
pingouin
transformers
xlsxwriter
hydra-core
pyreadstat
openpyxl